Refactor logic used to determine if a script should run or not
This works around an issue in Packer where, _very rarely_, the upload of
env vars fails. This is most likely due to a bug in the upstream WinRM
package.
The current set up is now configured in such a way that we will
generally get the desired behaviour even if the upload fails.