Workaround for Packer issue #8528: Install open-vm-tools using the preseed
Currently, Packer is unable to pick up when an instance changes IP
address following a reboot. Clearly, this leads to connectivity issues.

Installing the open-vm-tools package somehow works around the problem
and allows the instance to retain the IP address leased to it during
the install.

* Add the fuse and open-vm-tools packages to list of additional packages
  to be installed in the Debian preseed
* Remove options and scripts that installed and configured install of
  open-vm-tools and prepared instance for ease of use with shared
  folders

Note that the /mnt/hgfs directory is now automagically created and does
not need to be created manually