Revert "Workaround for Packer issue #8528: Install open-vm-tools using the preseed"
This reverts commit c618073ee1dd4d1f671394091d0248b99be00784.

Installing open-vm-tools using the preseed only worked once to resolve
the issues reported in #8528. This was pure coincidence.

The root cause of the issue is the fact that the DHCP client built into
the Debian 10 installer obtains an IP using the machines MAC address as
a unique identifier but, once built, the DHCP client installed into the
instance uses a DUID as it's unique identifier.

This means that the DHCP server treats the machine as two different
instances between when the installer is run and when the machine is
booted for the first time - effectively, the DHCP server issues one
address to the installer and another to the freshly booted instance.

This leads to problems connecting to the freshly built instance in
Packer - Packer only enumerates the IP address it should use to contact
the instance once at the very start of the build while the installer is
running.