Workaround for Packer issue #8528
See: https://github.com/hashicorp/packer/issues/8528

The ifupdown package in Debian 10 (ifupdown 0.8.35) is now configured to
call dhclient with a flag that sends the DUID as the host identifier
when obtaining an IP address from the DHCP server. This differs from the
installer which uses the MAC address as the identifier.

As a consequence, the DHCP server issues different IP addresses to the
installer and the newly built host. This causes issues for Packer as the
IP address used to contact the host is only enumerated when the
installer is running at the start of the build. Since the newly booted
host has a different IP address, Packer fails to establish a connection
to the host.

This commit fixes the issue by configuring the DHCP client in the newly
built host to use the MAC address - the same as is done by the
installer. Once Packer has established a connection, the change is
reverted to the default configuration shipped with the isc-dhcp-client
package.

Future versions of ifupdown will allow the user to configure if dhclient
is called with the flag to send the DUID to the DHCP server.