Fix for change in VirtualBox 7: Ensure the VM can reach the host network
A change in VirtualBox 7 means that, by default, VMs on a 'NAT' network
can no longer reach the host network. This meant that the VM could no
longer reach Packers HTTP server and as a consequence could not
retrieve its preseed file. This option now needs to be set to allow the
VM to connect to the host network.